requirements.txt
# --- Orchestration & Core ---
apache-airflow==2.10.0
requests==2.32.3

# --- Data Processing ---
duckdb==1.1.3
pandas==2.2.3
numpy==2.0.2

# --- Storage & Cloud Simulation ---
# boto3 is the "language" used to talk to S3 (and MinIO)
boto3==1.35.63
botocore==1.35.63

# --- Databases ---
# psycopg2-binary is the driver for PostgreSQL
psycopg2-binary==2.9.10
sqlalchemy==1.4.54

# --- Environment & Utils ---
python-dotenv==1.0.1
pyarrow==17.0.0

minio
duckdb
fsspec
s3fs